+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Rat's Mischief 1.1: When Rats'll Fly Update - 1.16.5
 - Added the Mask of Rat
     - Grants a bit of protection as well as all enchantments present on it to your rats when their owner has one equipped 
-    - TODO Makes the user share their status effects (positive and negative) with all their tamed 
+    - Makes the user share their status effects (positive and negative) with all their tamed rats
     - Can be repaired with Phantom Membrane
     - Will always drop when a player kills doctor4t
 - Tamed rats now drop what they hold upon dying